api/requirements.txt
gunicorn
django
djangorestframework
django-cors-headers
pandas
environs
numpy
openpyxl
xlrd
openai